bounding box calculation!
in the two "simpleZoom" examples, i've got the `zoomed` function
calculating the bounding box of the viewport in "canvas" coordinates,
which i think is exactly what i'll need to determine which slice of the
canvas should be drawn, and at what resolution.

whether the approach i currently have in mind will be performant or not
remains to be seen... i'm worried that "redrawing" everything manually
every tick is a terrible idea, so instead i'll have to try to use the
`.data()` d3 function in a smart way to get it to do the virtual DOM
thing... or maybe i move to three.js?  or HTML5 <canvas> first?  before
those, do it naively.  proof of concept, red-green-refactor (and yeah,
tests might be a good idea)